rename and fix gps2line

+++ R/gps2line.R
@@ -5,57 +5,51 @@
 #'
 #' @param input_filepath Character; location of exported GPS files
 #' @param output_filepath Character; location for export Line-data type
-#' @param fleet_data Character; path of fleet data associated with information of age, engine size,
+#' @param fleet_path Character; path of fleet data associated with information of age, engine size,
 #' weight gross total and shape_id's. This input associated with shape_id's allows the function to relate each
 #' shape_id to a vehicle, by creating columns of fleet specification in the output.
 #' Missing entry means no fleet allocation
 #' @param  overwrite Logical; overwrites files in the output_filepath
 #' @export
-gps_to_linestring <- function(input_filepath,output_filepath,fleet_data,overwrite = TRUE){
+gps2line <- function(input_filepath,output_filepath,fleet_path = NA,overwrite = TRUE){
   #
   # tests
-  # input_filepath = paste0("../../data/gps/",proj_cities$abrev_city)
-  # output_filepath <- paste0("../../data/gps_linestring/",proj_cities$abrev_city)
-  # fleet_data = readr::read_rds(paste0("../../data/fleet/",proj_cities$abrev_city,"/",proj_cities$abrev_city,".rds"))
-  #
-  # read fleet
-  #
-  tempd <- file.path(tempdir(), "fleet") # create tempr dir to save GTFS unzipped files
-  unlink(normalizePath(paste0(tempd, "/", dir(tempd)), mustWork = FALSE), recursive = TRUE) # clean tempfiles in that dir
-  utils::untar(tarfile = fleet_data,exdir = tempd) # untar files
-  f <- list.files(tempd)
-  fleet_data = data.table::fread(paste0(tempd,"/",f))
+  # input_filepath = "test_joao/gps/"
+  # output_filepath = "test_joao/lines/"
+  # fleet_path = "inst/extdata/cur_fleet.tar.xz"
+  # overwrite = TRUE
   #
   # list gps files in 'input_folder'
   #
-  gps_line <- list.files(input_filepath, recursive = FALSE,
-                      include.dirs = FALSE,full.names = TRUE)
-
-  output_name <- list.files(input_filepath, recursive = FALSE,
-                            include.dirs = FALSE,
-                            full.names = FALSE) 
+  gps_line <- list.files(input_filepath, recursive = FALSE)
   #
-  # check existing files in output_filepath
-  # files
+  # check existing files in output_filepath files
   #
-  check_files <- paste0(output_filepath,list.files(output_filepath))
-  check_files_names <- stringr::str_remove_all(list.files(output_filepath),".rds")
+  check_files <- paste0(output_filepath,list.files(output_filepath,recursive = TRUE))
+  #check_files_names <- gsub(".txt","",list.files(output_filepath))
   if(length(check_files) > 1 & overwrite == FALSE){
-    gps_line <- gps_line[gps_line %nin% check_files]
-    gps_line_names <- gps_line_names[gps_line_names %nin% check_files_names]
+    gps_line <- gps_line[gps_line %nin% list.files(output_filepath)]
   }
   #
   # fleet_data
   #
-  if(!missing(fleet_data)){  fleet_data[,hora_liberacao := data.table::as.ITime("00:00:00")] }
+  if(!is.na(fleet_path)){
+    tempd <- file.path(tempdir(), "fleet") # create tempr dir to save GTFS unzipped files
+    unlink(normalizePath(paste0(tempd, "/", dir(tempd)), mustWork = FALSE), recursive = TRUE) # clean tempfiles in that dir
+    utils::untar(tarfile = fleet_path,exdir = tempd) # untar files
+    f <- list.files(tempd)
+    # read fleet
+    fleet_data = data.table::fread(paste0(tempd,"/",f))
+    fleet_data[,hora_liberacao := data.table::as.ITime("00:00:00")]
+    }
   #
   # interation of all trip_id's
   #
-  lapply(1:length(gps_line),function(i){ # i = 2
+  lapply(1:length(gps_line),function(i){ # i = 1
     # message
-    message(paste0('shape_id: #', stringr::str_remove(output_name[i],".rds")," , ", i," out of ",length(gps_line)))
+    message(paste0('shape_id: #', gsub(".txt","",gps_line[i])," , ", i," out of ",length(gps_line)))
     # read
-    dt <- data.table::fread(paste0(gps_line[i]))
+    dt <- data.table::fread(paste0(input_filepath,gps_line[i]))
     dt[,id := 1:nrow(dt)]
     # stop division
     id0 <- c(1,which(!is.na(dt$stop_sequence)))
@@ -90,7 +84,7 @@ gps_to_linestring <- function(input_filepath,output_filepath,fleet_data,overwrit
     #
     # condition of fleet data
     #
-    if(!missing(fleet_data)){
+    if(!is.na(fleet_path)){
       real_fleet <- fleet_data[SHP %in% unique(dt3$shape_id) &
                                  hora_liberacao < dt3[,.SD[1]][,departure_time],]
       # sample and occupancy time
@@ -106,8 +100,8 @@ gps_to_linestring <- function(input_filepath,output_filepath,fleet_data,overwrit
     #
     # create output dir and save
     #
-    filepath <- paste0(output_filepath,"/",output_name[i])
-    data.table::fwrite(x = dt3,path = filepath)
+    filepath <- paste0(output_filepath,gps_line[i])
+    data.table::fwrite(x = dt3,file = filepath)
     return(NULL)
   })
   return(message('Files exported'))
